requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.1
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
idna==3.10
iyzipay==1.0.45
phonenumbers==9.0.2
pillow==10.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
python-whois==0.9.5
qrcode==7.4.2
reportlab==4.3.1
requests==2.31.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
whois==1.20240129.2
django-bootstrap5==23.4
django-widget-tweaks==1.5.0
django-cleanup==8.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
Pillow==10.2.0
dnspython==2.7.0
